==> 1_Traditional_ML/1_Preprocessing.ipynb <==
> RoyChao

## 1-0 Environments

# Python:
!python3 --version

# Pip:
!pip --version

# matplotlob, numpy, pandas, sklearn, scipy, seaborn
import matplotlib
print('The matplotlib version is {}'. format(matplotlib. __version__)) 
import numpy as np
print('The numpy version is {}.'.format(np.__version__))
import pandas as pd
print('The pandas version is {}.'.format(pd.__version__))
import sklearn
print('The scikit-learn version is {}.'.format(sklearn.__version__))
import scipy
print('The scipy version is {}.'.format(scipy.__version__))
import seaborn as sns
print('The seaborn version is {}.'.format(sns.__version__))

## 1-1 Environments Parameters

sns.set( style="ticks" )

## 1-2 Read CSV & Drop

df_raw = pd.read_csv("data/0_raw_elec.csv")
print(df_raw.head())

df_raw = df_raw[['日期', '淨尖峰供電能力(MW)', '尖峰負載(MW)', '工業用電(百萬度)', '民生用電(百萬度)']]
df_drop = df_raw.rename(columns=
{
    '日期' : 'date',
    '淨尖峰供電能力(MW)' : 'supply',
    '尖峰負載(MW)' : 'demand',
    '工業用電(百萬度)' : 'industry', 
    '民生用電(百萬度)' : 'civil'
})
print(df_drop.head())
print(df_drop.info())
print(df_drop.describe())

## 1-3 Make training dataframe

df_dataset = df_drop.copy()
df_dataset['OR'] = -1

for idx in range( len(df_dataset) - 1 ):
    df_dataset['OR'][idx] = df_dataset['supply'][idx + 1] - df_dataset['demand'][idx + 1]

df_dataset.drop(df_dataset.tail(1).index, inplace=True)     # drop last one
print( df_dataset.head() )
print( df_dataset.describe() )

df_dataset.to_csv("data/1_dataset_elec.csv")
