==> 1_Traditional_ML/2_Dataset_diff.ipynb <==
> Roychao

## 2-1 make different time period dataframe

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from matplotlib.pyplot import figure
import seaborn as sns
sns.set( style="ticks" )

df_dataset = pd.read_csv("data/1_dataset_elec.csv", index_col=0)
print(df_dataset.head())

## 2-1-2 remove outlier version

df_dataset.describe()

from sklearn.preprocessing import MinMaxScaler

min_max_scaler = MinMaxScaler()
df_tmp = df_dataset.copy()

min_max_scaler.fit(df_tmp[[
        'supply', 'demand', 'industry', 'civil']])
print(f'min: {min_max_scaler.data_min_}')
print(f'max: {min_max_scaler.data_max_}')

df_tmp[[
        'supply', 'demand', 'industry', 'civil']] = min_max_scaler.fit_transform(df_tmp[['supply', 'demand', 'industry', 'civil']])
df_tmp.drop(columns='date', inplace=True)

df_tmp.describe()

from scipy import stats
df_tmp[(np.abs(stats.zscore(df_tmp.drop(columns='OR'))) < 3).all(axis=1)]

from scipy import stats
df_outlier = df_dataset[(np.abs(stats.zscore(df_dataset.drop(columns=['OR', 'date']))) < 3).all(axis=1)]
# len( df_outlier )

min_max_scaler = MinMaxScaler()

min_max_scaler.fit(df_outlier[[
        'supply', 'demand', 'industry', 'civil'
        ]])
print(f'min: {min_max_scaler.data_min_}')
print(f'max: {min_max_scaler.data_max_}')

df_outlier[[
        'supply', 'demand', 'industry', 'civil'
        ]] = min_max_scaler.fit_transform(
            df_outlier[[
        'supply', 'demand', 'industry', 'civil'
        ]])


df_outlier.to_csv("data/3_dataset_outlier_elec.csv")

df_outlier.describe()

## 2-2 Make 2a_dataset_diffdays_elec.csv

from scipy.stats import skew
from scipy.stats import kurtosis

diff = diff

s_min = []
s_max = []
s_skew = []
s_kurt = []

d_min = []
d_max = []
d_skew = []
d_kurt = []

i_min = []
i_max = []
i_skew = []
i_kurt = []

c_min = []
c_max = []
c_skew = []
c_kurt = []

ans = []
date = []

for i in range( len(df_dataset) - diff ):
    idx = i + diff - 1
    
    s_min.append( min(df_dataset['supply'][idx - diff + 1 : idx + 1]) )
    s_max.append( max(df_dataset['supply'][idx - diff + 1 : idx + 1]) )
    s_skew.append( skew(df_dataset['supply'][idx - diff + 1 : idx + 1]) )
    s_kurt.append( kurtosis(df_dataset['supply'][idx - diff + 1 : idx + 1]) )
    
    d_min.append( min(df_dataset['demand'][idx - diff + 1 : idx + 1]) )
    d_max.append( max(df_dataset['demand'][idx - diff + 1 : idx + 1]) )
    d_skew.append( skew(df_dataset['demand'][idx - diff + 1 : idx + 1]) )
    d_kurt.append( kurtosis(df_dataset['demand'][idx - diff + 1 : idx + 1]) )

    i_min.append( min(df_dataset['industry'][idx - diff + 1 : idx + 1]) )
    i_max.append( max(df_dataset['industry'][idx - diff + 1 : idx + 1]) )
    i_skew.append( skew(df_dataset['industry'][idx - diff + 1 : idx + 1]) )
    i_kurt.append( kurtosis(df_dataset['industry'][idx - diff + 1 : idx + 1]) )

    c_min.append( min(df_dataset['civil'][idx - diff + 1 : idx + 1]) )
    c_max.append( max(df_dataset['civil'][idx - diff + 1 : idx + 1]) )
    c_skew.append( skew(df_dataset['civil'][idx - diff + 1 : idx + 1]) )
    c_kurt.append( kurtosis(df_dataset['civil'][idx - diff + 1 : idx + 1]) )

    ans.append( df_dataset['OR'][idx] )
    date.append( df_dataset['date'][idx] )

df_2 = pd.DataFrame(
    list(zip(
        ans, date,
        s_min, s_max, s_skew, s_kurt,
        d_min, d_max, d_skew, d_kurt,
        i_min, i_max, i_skew, i_kurt,
        c_min, c_max, c_skew, c_kurt)),
    columns =[
        'OR', 'date',
        's_min', 's_max', 's_skew', 's_kurt',
        'd_min', 'd_max', 'd_skew', 'd_kurt',
        'i_min', 'i_max', 'i_skew', 'i_kurt',
        'c_min', 'c_max', 'c_skew', 'c_kurt',]
)

df_2.to_csv("data/2_df2_elec.csv")

## 2-3 diff Make 2a_dataset_5days_elec

from scipy.stats import skew
from scipy.stats import kurtosis

diff = 5

s_min = []
s_max = []
s_skew = []
s_kurt = []

d_min = []
d_max = []
d_skew = []
d_kurt = []

i_min = []
i_max = []
i_skew = []
i_kurt = []

c_min = []
c_max = []
c_skew = []
c_kurt = []

ans = []
date = []

for i in range( len(df_dataset) - diff ):
    idx = i + diff - 1
    
    s_min.append( min(df_dataset['supply'][idx - diff + 1 : idx + 1]) )
    s_max.append( max(df_dataset['supply'][idx - diff + 1 : idx + 1]) )
    s_skew.append( skew(df_dataset['supply'][idx - diff + 1 : idx + 1]) )
    s_kurt.append( kurtosis(df_dataset['supply'][idx - diff + 1 : idx + 1]) )
    
    d_min.append( min(df_dataset['demand'][idx - diff + 1 : idx + 1]) )
    d_max.append( max(df_dataset['demand'][idx - diff + 1 : idx + 1]) )
    d_skew.append( skew(df_dataset['demand'][idx - diff + 1 : idx + 1]) )
    d_kurt.append( kurtosis(df_dataset['demand'][idx - diff + 1 : idx + 1]) )

    i_min.append( min(df_dataset['industry'][idx - diff + 1 : idx + 1]) )
    i_max.append( max(df_dataset['industry'][idx - diff + 1 : idx + 1]) )
    i_skew.append( skew(df_dataset['industry'][idx - diff + 1 : idx + 1]) )
    i_kurt.append( kurtosis(df_dataset['industry'][idx - diff + 1 : idx + 1]) )

    c_min.append( min(df_dataset['civil'][idx - diff + 1 : idx + 1]) )
    c_max.append( max(df_dataset['civil'][idx - diff + 1 : idx + 1]) )
    c_skew.append( skew(df_dataset['civil'][idx - diff + 1 : idx + 1]) )
    c_kurt.append( kurtosis(df_dataset['civil'][idx - diff + 1 : idx + 1]) )

    ans.append( df_dataset['OR'][idx] )
    date.append( df_dataset['date'][idx] )

df_2 = pd.DataFrame(
    list(zip(
        ans, date,
        s_min, s_max, s_skew, s_kurt,
        d_min, d_max, d_skew, d_kurt,
        i_min, i_max, i_skew, i_kurt,
        c_min, c_max, c_skew, c_kurt)),
    columns =[
        'OR', 'date',
        's_min', 's_max', 's_skew', 's_kurt',
        'd_min', 'd_max', 'd_skew', 'd_kurt',
        'i_min', 'i_max', 'i_skew', 'i_kurt',
        'c_min', 'c_max', 'c_skew', 'c_kurt',]
)

df_2.to_csv("data/2_df2_5day_elec.csv")

## 2-4 Make 2a_dataset_7days_elec

from scipy.stats import skew
from scipy.stats import kurtosis

diff = 7

s_min = []
s_max = []
s_skew = []
s_kurt = []

d_min = []
d_max = []
d_skew = []
d_kurt = []

i_min = []
i_max = []
i_skew = []
i_kurt = []

c_min = []
c_max = []
c_skew = []
c_kurt = []

ans = []
date = []

for i in range( len(df_dataset) - diff ):
    idx = i + diff - 1
    
    s_min.append( min(df_dataset['supply'][idx - diff + 1 : idx + 1]) )
    s_max.append( max(df_dataset['supply'][idx - diff + 1 : idx + 1]) )
    s_skew.append( skew(df_dataset['supply'][idx - diff + 1 : idx + 1]) )
    s_kurt.append( kurtosis(df_dataset['supply'][idx - diff + 1 : idx + 1]) )
    
    d_min.append( min(df_dataset['demand'][idx - diff + 1 : idx + 1]) )
    d_max.append( max(df_dataset['demand'][idx - diff + 1 : idx + 1]) )
    d_skew.append( skew(df_dataset['demand'][idx - diff + 1 : idx + 1]) )
    d_kurt.append( kurtosis(df_dataset['demand'][idx - diff + 1 : idx + 1]) )

    i_min.append( min(df_dataset['industry'][idx - diff + 1 : idx + 1]) )
    i_max.append( max(df_dataset['industry'][idx - diff + 1 : idx + 1]) )
    i_skew.append( skew(df_dataset['industry'][idx - diff + 1 : idx + 1]) )
    i_kurt.append( kurtosis(df_dataset['industry'][idx - diff + 1 : idx + 1]) )

    c_min.append( min(df_dataset['civil'][idx - diff + 1 : idx + 1]) )
    c_max.append( max(df_dataset['civil'][idx - diff + 1 : idx + 1]) )
    c_skew.append( skew(df_dataset['civil'][idx - diff + 1 : idx + 1]) )
    c_kurt.append( kurtosis(df_dataset['civil'][idx - diff + 1 : idx + 1]) )

    ans.append( df_dataset['OR'][idx] )
    date.append( df_dataset['date'][idx] )

df_2 = pd.DataFrame(
    list(zip(
        ans, date,
        s_min, s_max, s_skew, s_kurt,
        d_min, d_max, d_skew, d_kurt,
        i_min, i_max, i_skew, i_kurt,
        c_min, c_max, c_skew, c_kurt)),
    columns =[
        'OR', 'date',
        's_min', 's_max', 's_skew', 's_kurt',
        'd_min', 'd_max', 'd_skew', 'd_kurt',
        'i_min', 'i_max', 'i_skew', 'i_kurt',
        'c_min', 'c_max', 'c_skew', 'c_kurt',]
)

df_2.to_csv("data/2_df2_7day_elec.csv")

## 2-5 Make 2a_dataset_14days_elec

from scipy.stats import skew
from scipy.stats import kurtosis

diff = 14

s_min = []
s_max = []
s_skew = []
s_kurt = []

d_min = []
d_max = []
d_skew = []
d_kurt = []

i_min = []
i_max = []
i_skew = []
i_kurt = []

c_min = []
c_max = []
c_skew = []
c_kurt = []

ans = []
date = []

for i in range( len(df_dataset) - diff ):
    idx = i + diff - 1
    
    s_min.append( min(df_dataset['supply'][idx - diff + 1 : idx + 1]) )
    s_max.append( max(df_dataset['supply'][idx - diff + 1 : idx + 1]) )
    s_skew.append( skew(df_dataset['supply'][idx - diff + 1 : idx + 1]) )
    s_kurt.append( kurtosis(df_dataset['supply'][idx - diff + 1 : idx + 1]) )
    
    d_min.append( min(df_dataset['demand'][idx - diff + 1 : idx + 1]) )
    d_max.append( max(df_dataset['demand'][idx - diff + 1 : idx + 1]) )
    d_skew.append( skew(df_dataset['demand'][idx - diff + 1 : idx + 1]) )
    d_kurt.append( kurtosis(df_dataset['demand'][idx - diff + 1 : idx + 1]) )

    i_min.append( min(df_dataset['industry'][idx - diff + 1 : idx + 1]) )
    i_max.append( max(df_dataset['industry'][idx - diff + 1 : idx + 1]) )
    i_skew.append( skew(df_dataset['industry'][idx - diff + 1 : idx + 1]) )
    i_kurt.append( kurtosis(df_dataset['industry'][idx - diff + 1 : idx + 1]) )

    c_min.append( min(df_dataset['civil'][idx - diff + 1 : idx + 1]) )
    c_max.append( max(df_dataset['civil'][idx - diff + 1 : idx + 1]) )
    c_skew.append( skew(df_dataset['civil'][idx - diff + 1 : idx + 1]) )
    c_kurt.append( kurtosis(df_dataset['civil'][idx - diff + 1 : idx + 1]) )

    ans.append( df_dataset['OR'][idx] )
    date.append( df_dataset['date'][idx] )

df_2 = pd.DataFrame(
    list(zip(
        ans, date,
        s_min, s_max, s_skew, s_kurt,
        d_min, d_max, d_skew, d_kurt,
        i_min, i_max, i_skew, i_kurt,
        c_min, c_max, c_skew, c_kurt)),
    columns =[
        'OR', 'date',
        's_min', 's_max', 's_skew', 's_kurt',
        'd_min', 'd_max', 'd_skew', 'd_kurt',
        'i_min', 'i_max', 'i_skew', 'i_kurt',
        'c_min', 'c_max', 'c_skew', 'c_kurt',]
)

df_2.to_csv("data/2_df2_14day_elec.csv")
