==> 1_Traditional_ML/4_DL_analysis.ipynb <==
import glob
import os
import pandas as pd
import numpy as np
from sklearn.metrics import mean_squared_error, mean_absolute_error

folder_path = "/Users/blueshiner/git/3_Collections/19_DSAI/HW0/1_codes/src/results/*.csv"

# All files and directories ending with .txt and that don't begin with a dot:
# print(glob.glob(folder_path)) 

for item in glob.glob(folder_path):
    #df = pd.read_csv(item)
    pred = np.loadtxt(item, delimiter=',')

    print("----------------------------------------------")
    print(f"File Name: {os.path.basename(item)}")
    mse = mean_squared_error( pred, [3032,2301,3167,2963,2549,3464,4700,3346,3087,3111]) / 10
    mae = mean_absolute_error( pred, [3032,2301,3167,2963,2549,3464,4700,3346,3087,3111]) / 10
    print(f"    MSE: {mse}, MAE: {mae}")
